feat: add page-specific metadata and OG tags to core pages #282
